Make namei() match the unveils against the requested Capsicum rights.
Drop the extra namei flags that were used just for vn_open() and co.

Most namei() calls already pass the specific rights that they expect
thanks to Capsicum, so take advantage of that instead of trying to
change tons of call sites.

Change some calls that used a "naked" NDINIT() to use NDINIT_ATRIGHTS()
with appropriate rights.  These were from syscalls that Capsicum didn't
need to deal with.  There are still some NDINIT()s that probably should
be fixed.  Some are root-only but should be fixed eventually too to be
able to pledge() root processes.