Reject all namei() calls with no capability rights passed.
Unveil must be starting to be pretty well locked-down now.

Add a ni_intflags field to namei for "internal flags" and start using
this instead of using up the very last cn_flags.  ni_lcf is too
ephemeral for this purpose.

readlink(2) was not requiring any capability rights.  Modify it to use
CAP_FSTAT.