Move permission "inheritance" handling in the kernel.
Trying to make the userland wrapper do the right thing was a PITA and I
wasn't able to simplify what's in the kernel as much as I had hoped
anyway.  It's simpler overall now.

Adding unveils for intermediate path components isn't done anymore.
Will have the kernel do it later.