Make pledge() work after unveil(NULL, NULL).
Previously, pledge() had to be done before the final unveil.  But some
OpenBSD code does it in the opposite order.

If a pledge() hasn't already been made, pre-unveil paths for all
possible promises before freezing the unveils.  This works fine with
inherited pledge unveils.

Rearrange flags handling a bit.  Make it clearer that the FREEZE
operation applies to both roles (unlike SWEEP).  This is going to need
better comments.