Bunch of reorganizations and cleanups.
Drop unveil caching completely for now.  The filedesc is difficult to
deal with and it gets in the way of other changes.  There might be a
simpler way to do this (that would also make this whole unveil/pledge
support a less intrusive diff).

Make the unveil hooks in lookup() more coherent.

Make more stuff internal to sys_unveil.c.