Add a "route" promise/sysfil.
Require it to read the net.routetable sysctl instead of just "net".

Also enable it for the "dns" promise because getaddrinfo() needs it.

There's a bunch of ioctls that also should be allowed under "route" but
not sure yet how to go about that.  And they probably shouldn't all be
allowed just for DNS resolution.