Add alternative process visibility restriction sysfils.
Add promise variants for each.

Make veilbox default to just (direct) child process visibility.
firefox/chromium apparently can work fine with just that.  Add a process
group mode.  Add a shell spawning mode with whole session visibility.
This should spawn the shell in a new session but it doesn't yet (it'll
probably need the parent process to stick around and wait on it).