Allow to chdir() to intermediate directories that aren't directly unveiled.
It makes some programs work better.  In particular programs that try to
chdir to / while daemonizing.

OpenBSD doesn't allow this, but it also stopped allowing stat() on the
intermediate directories since they fixed a lot of programs to not
require it.

There's a way to toggle this with unveilctl(), but it's not exposed.