Fix running test already sandboxed.
kyua runs the test in a $TMPDIR subdirectory and the "tmpdir" pledge()
slot unveils get (recursively) combined with the test's unveil() slot
unveils and makes the whole thing accessible with too much permissions.

Re-pledging GCs and refreshes those unveils with the new $TMPDIR.

Apart from the "tmpdir" promise being much too permissive compared to
OpenBSD (for now), this is probably correct behavior; unveil() generally
shouldn't be used without a pledge() except when testing or when it is
known that the pledge is up-to-date.