Enhance unveil "slots" support.
Instead of having slots directly attached to each "role" in a fixed
configuration, allow each slot to be dynamically toggled for each role.

Slots are no longer individually inherited on execution; only the
"frozen" permissions matter.  Each executed process starts with more of
a clean slate and must re-initialize all of the slots it intends to use.
This makes it much less risky to retain permissions unintentionally.

The extra slots could be used to let the user have multiple sets of
unveils that don't cancel each others out (and veilbox could make use of
that), but there should be a good way to avoid slot number conflicts.