Add new "wrapper" mode to veilbox
Spawn sandboxed program in a child process and wait on it to perform
cleanups.  Use it to give each sandbox its own independent tmpdir.

Drop previous per-UID shared "blind" tmpdir support and just rely on the
"tmppath" promise when not in wrapper mode.  Which can be very insecure
depending on what's in tmpdir.

Wrapper mode should probably be made the default.