Add X11 wrapper mode
Create a temporary XAUTHORITY file with a cookie for the display.
Supports creating an "untrusted" cookie.

Use same flags (-X and -Y) as ssh(1).

There's SOMETHING, a so-called X11 "authentication spoofing" mechanism
in ssh that seems to interfere with cookie generation and makes `ssh -X
host veilbox -X cmd ...` not work.  But `ssh -X host veilbox -Y cmd ...`
works and should be just as secure (since the X11 cookie from SSH is
untrusted to begin with).