Change CAP_CONNECTAT sysfil/unveil permissions
Change requirement from "cpath"/"c" to "wpath"/"w".

This doesn't fit very well with the existing permissions.  On the one
hand, the ability to connect to a local domain socket can be very
dangerous and could allow to escape the sandbox depending on what's on
the other side.  On the other hand, it should be possible to grant a
sandboxed application the ability to connect to a specific socket
without giving too many permissions on the underlying path.

On OpenBSD, local domain socket aren't checked against unveils (IIUC).