Plug O_EMPTY_PATH security hole.  Hopefully.
Using file flags for stuff like that is wrong.  It changes behavior too
much compared to non-unveil mode and will probably cause problems.

Not sure what's a better way to do it.  Opened non-directory files can't
be re-checked against the current unveils because it might not be known
what directory(ies) they're in.