Make openpty()/forkpty() work with SYSFIL_TTY
Found a simple trick which seems to be good enough for a lot of programs.

After creating a pty device, the slave device will be hidden again the
first time anything changes the process' ucred.  Which a lot of programs
are OK with as long as /dev/tty still works.