Overhaul config tags/sections handling
Now a lot smarter about not re-processing sections needlessly.  Remove
arbitrary tags limit.

Probably needs more testing though.

Not sure sticking with this simple line-based format was a good idea but
I think it'll be able to do the job well enough with a few extensions to
the tags system.