Better way to protect shared /tmp
This should be moved to the config file, but there would have to be a
way to substitute the current UID into the paths.

This shouldn't be relied upon too much, but is better than nothing when
having to unveil too much to make something work.

Add function to reinitialize curtain(3) as a whole which wasn't too
appropriate for this but might be useful eventually.