Rework on-self/on-exec and "unrestricted" handling
Instead of relying on curtainctl(2) top-level flags to select which role
to affect, add a new way to represent an "unrestricted" role and always
have all roles fully specified with each call.

This allows to switch back-and-forth between soft restricted and
unrestricted mode as long as the hard permissions allow it.