Convert to a MAC module, lots of changes
Move as much of the code as possible to a separate MAC module.  Access
checks now done with MAC hooks and callbacks.  Curtains are pointed to
by MAC labels.

There's still some clutter that could probably be moved to the module
some way.  The sysfil stuff is intended to remain in the kernel.

Many VFS checks are now done a bit after the namei() and this lets the
calling code handle special cases and return the right errnos and
overall this makes things less weird for sandboxed programs.

Also namei() will now completely refuse to traverse non-unveiled path
components.  There's a new unveil permission for intermediate components
that only allows to traverse them (which is done by default).

Some of the new MAC hooks should probably be rethought (or maybe they
should just be separate hooks and this doesn't need to be in the MAC
framework at all) and the per-thread "tracker" handling is pretty messy.
The unveil code could use some more cleanups before the rest can be.
Ideally unveils could be stored directly in the curtain with the rest
but there are problems with that.

This probably introduced some new issues, but so far seems to work well
with sandboxed browsers, desktop apps, samba, and `make world`.