Enforce curtain hierarchy barriers for posixshm accesses
There could be separate barriers for different types of objects, but
this will do for now.

By itself this is not sufficient to secure POSIX shm/sem objects because
the paths need to be restricted as well, not just access to the objects
they point to.  But this could be a good way to deal with SysV IPC
objects.  And this could be something that the curtain module generally
does for all object types.