curtain: overhaul unveils handling
Attach some of the unveil bookkeeping to curtains.  This allows to get a
snapshot of the unveil tree without locking and allows to get rid of
some overhead that was added to syscall entry.  Replace kludge used for
file creation tracking with a less bad kludge.  It could be cleaned up
and simplified some more but this seems to be working well.