curtain: Improve PTY wrapping
Stop opening /dev/tty in partial mode.  This might be unexpected and
lead to unintended TTY sharing.  Try to get a TTY from any of the 3
standard FDs instead.

Try to make ^Z/fg work while PTY wrapping.  Both when STDIN is wrapped
and unwrapped.  In both cases, suspend the parent in reaction to the
child being suspended.

This is gonna need some more testing... this is a pain to get right.