curtain(1): Don't switch TTY to raw mode when there are redirections
When the redirections are pipes there can be concurrent use of the TTY
and the other programs won't work well.

On the other hand the wrapped program won't get real raw mode behavior
when it expects it (like getting a password from /dev/tty) if there are
redirections.

There might be no way to get both correctly at the same time.