curtain: sysctl hierarchy handling
I wanted to make curtain items refer to sysctl_oids directly, but seems
like there's no way to hold on to sysctl_oids after the sysctl lock is
released.  They can be allocated from memory mappings that get nuked
when modules are unloaded IIUC.

Attach "shadow" objects to sysctl_oids that can be used as an indirect
handle that can outlive them.  All that's needed is to be able to figure
out their parent links.