curtain: Replace checks for SYSFIL_DEFAULT with more specific checks
SYSFIL_DEFAULT probably shouldn't ever be used directly since Capsicum
does not drop it and SYSFIL_DEFAULT usually represents operations that
would be unsafe with Capsicum too (even though the syscall was disabled
in this case).  CRED_IN_VFS_VEILED_MODE() is an exception to that.

Use SYSFIL_FULL instead.  Fix curtain's handling of unknown sysfils by
treating them as if they were associated with CURTAINABL_DEFAULT (it was
a bit SIGKILL-happy otherwise).