curtain(1/5): Change default application level handling
Stop trying to restrict applications so much when not killing them on
restriction violations (-k option).  The problem is that unexpected
restrictions can make applications behave erratically and this could
then lead to security issues (as well as other problems like corrupting
files).  But if the application is killed on violation then it shouldn't
have time to do harm (other than by not working).

Reorganize some profiles around a "loose" and "strict" mode.  Strict
mode is only enabled with -k.  Add some non-killing restrictions to
strict mode to (try to) keep applications alive but it's probably
missing some.

Strict mode restrictions are going to be harder to figure out (and
organize).

Loose (non -k) mode will default to level 5 which is permissive enough
for most applications but it could still cause security issues in some
cases.

Maybe the default should be raised or be some kind of hybrid mode.