curtain: Add "simple" sandbox API to libc
Tries to load libcurtain and apply configuration for a tag.

The idea is that it should be possible to add very simple sandboxing
"hooks" to applications without having to link them against libcurtain.

With a small patch, firefox can have per-process isolation with it
(which can be on top of sandboxing firefox as a whole).