curtain: Add some protection to "passdir" ability
It's a dangerous ability but some programs (e.g. chrome, wine) seem to
need it.

Problem is that an untrusted sandboxed process could send a directory FD
to an untrusted but unsandboxed (or sandboxed with less restrictive
unveils) process that could then traverse it with less restrictions than
the sender process.  Processes from the same user should have proper
restrictions, but processes from other users may not.  Filesystem-level
permissions usually protect users from one another, but with an
untrusted process sending directory FDs it becomes possible to bypass
the restrictions on parent directories.

This forbids receiving directory FDs with different UIDs/GIDs sent by
processes that are lower down in the barriers hierarchy.