curtain: Drop process sysfil checks, change to debug ability
Stop checking for sysfils in the p_can*() functions.  The checks done by
the MAC handlers are enough.

p_candebug() is now always allowed on the current process.  The "debug"
ability is only needed for operating on other processes.  Makes the
"aslr" ability usable without "debug".

Add separate "trace" ability for ptrace(2)/ktrace(2) so that "debug" can
be enabled with a bit less risk.