curtain(4): Unveil kludge for device-specific TTY files
Some programs want to access their controlling tty through the
device-sepcific file (e.g. /dev/pts/*, /dev/ttyv*, etc), not just the
generic /dev/tty one.

One way was to run them in a PTY created within the curtain (since there
was already a kludge for that) but now it won't be necessary.